Move the Interface Definition for /booking/add to the Common Project.
