+++ Backend/common/src/main/java/com/bikenest/common/interfaces/bikenest/ExampleInterface.java
@@ -1,4 +1,4 @@
-package com.bikenest.common;
+package com.bikenest.common.interfaces.bikenest;
 
 public class ExampleInterface {
     private String test;
