+++ Backend/service-booking/src/main/java/com/bikenest/servicebooking/DB/Reservation.java
@@ -1,6 +1,6 @@
 package com.bikenest.servicebooking.DB;
 
-import com.bikenest.servicebooking.Reservation.NewReservationPOJO;
+import com.bikenest.common.interfaces.booking.AddReservationInterface;
 
 import javax.persistence.*;
 import java.time.Instant;
@@ -29,13 +29,14 @@ public class Reservation {
     @Temporal(TemporalType.TIMESTAMP)
     private Date ActualEndDateTime;   //When was the Bike actually taken from the Bikenest?
 
-    public static Reservation FromNewReservation(NewReservationPOJO newReservationPOJO){
+    public static Reservation FromNewReservation(Integer userId, AddReservationInterface newReservationPOJO){
         Reservation result = new Reservation();
         result.setActualEndDateTime(null);
         result.setActualStartDateTime(null);
         result.setStartDateTime(newReservationPOJO.getStartDateTime());
         result.setEndDateTime(newReservationPOJO.getEndDateTime());
         result.setBikenestId(newReservationPOJO.getBikenestId());
+        result.setUserId(userId);
         return result;
     }
